+++ src/main/java/ru/job4j/cars/repository/owner/OwnerRepository.java
@@ -8,9 +8,9 @@ import java.util.Optional;
 public interface OwnerRepository {
     Optional<Owner> create(Owner owner);
     boolean update(Owner owner);
-    boolean delete(int id);
+    boolean delete(Long id);
     List<Owner> findAllOrderById();
-    Optional<Owner> findById(int id);
+    Optional<Owner> findById(Long id);
     List<Owner> findByLikeName(String name);
     Optional<Owner> findByName(String name);
 }
